Fixed self collision bug,
When pressing two directions at once, the snake was allowed to move in a direction it should not be able to. I added a "pending direction" variable to prevent this from happening. Also, my directions were in reverse thanks to using the wrong keys for each arrow on the keyboard; fixed that as well.